+++ CHANGELOG.md
@@ -1,6 +1,4 @@
 # Cynosure `@VERSION@` `1.20.1` `Fabric & Forge`
 
 ### Changelog:
-- Fixed event buses
-- Fixed AccessWideners and Transformers
-- made global storage try to use the game folder if the main files fail
+- ServerChatEvent for messages sent to the player on the Server!